requirements.txt
python-dotenv
aiogram #python package helps you to connect with telegram
openai 
